requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.21
pydantic==1.10.9
python-dotenv==1.0.0
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.8
pymysql==1.0.3
